nebula/requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-import-export==3.3.1
django-cors-headers==4.2.0